fix(onboarding): show 'Run /kata to get started' on banner, remove 'Skip for now' option
On first launch with an unconfigured project, the startup banner now
shows 'Run /kata to get started.' below the web search status line,
so users know how to begin without guessing.

Removed the 'Skip for now' wizard option which disabled /kata for the
entire session. The wizard now has two choices: 'Set up Kata' and
'Not yet' (the auto-appended escape hatch from showNextAction).
